pyproject.toml
[project]
name = "cvlint"
version = "0.1.0"
description = ""
authors = [
    {name = "Dalton Luce",email = "daltonluce42@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
    "pypdf (>=6.0.0,<7.0.0)",
    "pyspellchecker (>=0.8.3,<0.9.0)",
    "pymupdf (>=1.26.4,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pdfminer-six (>=20250506,<20250507)",
    "typer[all] (>=0.12.0,<1.0.0)",
    "rich (>=13.0.0,<14.0.0)"
]

[project.scripts]
cvlint = "cvlint.cli:app"

[tool.poetry]
packages = [{include = "cvlint", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "reportlab (>=4.0.0,<5.0.0)",
    "black (>=25.9.0,<26.0.0)"
]
